pyproject.toml
[project]
name = "QBot"
version = "0.1.0"
description = "QQBot"
readme = "README.md"
authors = [
    { name = "dudu", email = "dudu7615@qq.com" },
]
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pyyaml>=6.0.3",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
